requirements.txt
fastapi
requests
pydantic
python-dotenv
openai
supabase
streamlit-cookies-manager
streamlit-javascript==0.1.5
urllib3==1.26.18
